+++ src/test/kotlin/com/github/ttyniwa/intellij/plugin/align/AlignDemoCode.kt
@@ -1,7 +1,7 @@
 package com.github.ttyniwa.intellij.plugin.align
 
 class AlignDemoCode {
-    fun hoge() {
+    fun alignAroundCursor() {
         var aaa = 1
         var b = 2
         b += 2
@@ -17,5 +17,15 @@ class AlignDemoCode {
                 "hello world", "yes",   // comment
                 "good bye"
         )
+    }
+
+    fun selectAndAlign() {
+        val notAligned = 1
+        var aaa = 1
+        var b = 2
+
+        aaa += 123
+        b += 2
+        val notAligned2 = 1
     }
 }